@@ -6,6 +6,7 @@ use App\Repository\UserChiefRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use JetBrains\PhpStorm\Pure;
 
 /**
  * @ORM\Entity(repositoryClass=UserChiefRepository::class)
@@ -52,12 +53,12 @@ class UserChief
     /**
      * @ORM\OneToMany(targetEntity=Booking::class, mappedBy="chief", orphanRemoval=true)
      */
-    private ArrayCollection $bookings;
+    private Collection $bookings;
 
     /**
      * @ORM\OneToMany(targetEntity=Live::class, mappedBy="userChief", orphanRemoval=true)
      */
-    private ArrayCollection $lives;
+    private Collection $lives;
 
     /**
      * @ORM\Column(type="string", length=255)
@@ -67,7 +68,7 @@ class UserChief
     /**
      * @ORM\Column(type="json")
      */
-    private array $equipments = [];
+    public array $equipments = [];
 
     /**
      * @ORM\Column(type="string", length=255, nullable=true)
@@ -79,7 +80,7 @@ class UserChief
      */
     private ?string $description;
 
-    public function __construct()
+    #[Pure] public function __construct()
     {
         $this->bookings = new ArrayCollection();
         $this->lives = new ArrayCollection();
